feat: Project stage 1.
The project's stage 1 consists of getting movies' images posters and shown them as a list. The list must present the most popular movies or the top rated movies. When the user clicks on a movie from the list, the application must show detailed informations about that movie. Moreover, the application must be designed for tablets and handheld phones.